# openapi_ 3.0.3.yml
openapi: 3.0.3
info:
  title: "Social Media Apps"
  description: "Project 1 Group 5 Social Media Apps"
  version: 1.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/aziz-wahyudin/Social-Media-Apps/1.0.0
  - url: "http://18.143.74.57:8000" # perlu diubah menyesuaikan dengan host domain
    description: "link project"
...
paths:
...
  /users:
    post:
      tags:
        - "Users"
      summary: "Registration feature for user"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/RegistrationObject"
            example:
              username: "aziz123"
              email: "azizw@duck.com"
              password: "123456789" 
      responses:
        '201':
          description: "success create account"
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/ResponseObject"
              example:
                message: "success register account"
                data:
                  id_user: 2
                  username: "aziz123"
                  email: "azizw@duck.com"
                  password: "123456789"
        '400':
          $ref: "#/components/responses/400"
        '500':
          $ref: "#/components/responses/500"
    put:
      tags:
        - "Users"
      summary: "Update user profile by account owner"
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              $ref: "#/components/schemas/RequestUpdate"
            example:
              username: "aziz123"
              email: "azizw@duck.com"
              password: "123456789"
              photo: aziz123.jpg
              bio: "oplover"
      responses:
        '200':
          description: "success update user"
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/ResponseObject"
              example:
                message: "success update user"
                data:
                  id_user: 2
                  username: "aziz123"
                  email: "azizw@duck.com"
                  password: "123456789"
                  photo: aziz123.jpg  
                  bio: "oplover"
        '400':
          $ref: "#/components/responses/400"
        '500':
          $ref: "#/components/responses/500"  
    delete:
      tags:
        - "Users"
      security:
      - bearerAuth: []
      summary: Delete by id
      responses:
        200: 
          description: "Success deactive account"
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/ResponseObject'
              example:
                message: "Success deactive account"
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
    get:
      tags:
        - "Users"
      summary: "shows login user profile"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Success get user profile"
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/ResponseArray"
              example:
                message: "success get user profile"
                data:
                  id_user: 2
                  username: "aziz123"
                  email: "azizw@duck.com"
                  password: "123456789"
                  photo: aziz123.jpg  
                  bio: "oplover"
                  post: 
                    - id_post: 1
                      content: "#spoileronepiece1064"
                      created_at: "2022-10-17 10:24:17"
                      owner: aziz123
                    - id_post: 2
                      content: "#spoilerkaguya-sama"
                      created_at: "2022-10-18 10:24:17"
                      owner: aziz123
        '400':
          $ref: "#/components/responses/400"
        '500':
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegistrationObject:
      type: object
      properties:
        id_user: 
          type: integer
        username:
          type: string
        email: 
          type: string
        password:
          type: string
          minLength: 4
        photo:
          type: string
    RequestUpdate:
      type: object
      properties:
        username:
          type: string
        email: 
          type: string
        password:
          type: string
        photo:
          type: string
        bio: 
          type: string
...
    ResponseObject:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
    ResponseArray:
      type: object
      properties:
        message: 
          type: string
        data:
          type: array
          items:
            type: object
...
  responses:
...
    '400':
      description: "error from user"
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
          example:
            message: "cannot process data, invalid input from user"
...
    '404':
      description: "The requested resource was not found."
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
          example:
              message: "Requested resource was not found"
    '500':
      description: "error from server"
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
          example:
            message: "cannot process data, something wrong on server"